cloud-functions/check-and-retry-clips/requirements.txt
functions-framework==3.*
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1
google-auth==2.22.0
requests==2.31.0
google-genai>=0.8.0
